Move IPC interface types to its own directory
